Add files via upload
"Who Did?" is a game developed using Java as part of our Software Engineering class project. The project required us to come up with a concept and build it from start to finish, following all necessary steps. This included ideation, planning, design, implementation, and testing. We also prepared detailed documentation and presentations to showcase our progress and final product. The culmination of the project involved demonstrating the game to our professor, highlighting both the technical aspects and the user experience. This project allowed us to apply theoretical knowledge in a practical setting and gain hands-on experience in the entire software development lifecycle.

the project itself is located here --- https://github.com/viktorshkrivani/whodid